Fake login with Html, Css, Jquery
Creation of fake login that basically does a validation
of username and password.

when validate add <span> under input label showing the error
if the username or password is empty or the username is not
a email.

if the user or password is wrong show a modal window with
error or sucess.